+++ frontend/src/app/modules/page/parts/model/parts.model.ts
@@ -34,7 +34,8 @@ export enum QualityType {
 
 export enum SemanticDataModel {
   BATCH= 'BATCH',
-  SERIALPARTTYPIZATION = 'SERIALPARTTYPIZATION'
+  SERIALPARTTYPIZATION = 'SERIALPARTTYPIZATION',
+  PARTASPLANNED = 'PARTASPLANNED'
 }
 
 export interface Part {